Refactor: Moving ConversationInfo to server module
