+++ openhands/server/data_models/conversation_info.py
@@ -6,7 +6,10 @@ from openhands.storage.data_models.conversation_status import ConversationStatus
 
 @dataclass
 class ConversationInfo:
-    """Information about a conversation"""
+    """
+    Information about a conversation. This combines conversation metadata with
+    information on whetehr a conversation is currently running
+    """
 
     conversation_id: str
     title: str
