+++ tests/unit/test_conversation.py
@@ -6,15 +6,15 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from openhands.runtime.impl.docker.docker_runtime import DockerRuntime
+from openhands.server.data_models.conversation_info import ConversationInfo
+from openhands.server.data_models.conversation_info_result_set import (
+    ConversationInfoResultSet,
+)
 from openhands.server.routes.manage_conversations import (
     delete_conversation,
     get_conversation,
     search_conversations,
     update_conversation,
-)
-from openhands.storage.data_models.conversation_info import ConversationInfo
-from openhands.storage.data_models.conversation_info_result_set import (
-    ConversationInfoResultSet,
 )
 from openhands.storage.data_models.conversation_status import ConversationStatus
 from openhands.storage.memory import InMemoryFileStore
